--- a/lightstep/sdk/internal/common.go
+++ b/lightstep/sdk/internal/common.go
@@ -61,43 +61,46 @@ func (rm *ResourceMap) Get(in *resource.Resource) pcommon.Resource {
 
 func CopyAttributes(dest pcommon.Map, src attribute.Set) {
 	for iter := src.Iter(); iter.Next(); {
-		inA := iter.Attribute()
-		key := string(inA.Key)
-		switch inA.Value.Type() {
-		case attribute.BOOL:
-			dest.PutBool(key, inA.Value.AsBool())
-		case attribute.INT64:
-			dest.PutInt(key, inA.Value.AsInt64())
-		case attribute.FLOAT64:
-			dest.PutDouble(key, inA.Value.AsFloat64())
-		case attribute.STRING:
-			dest.PutStr(key, inA.Value.AsString())
-		case attribute.BOOLSLICE:
-			sl := dest.PutEmptySlice(key)
-			sl.EnsureCapacity(len(inA.Value.AsBoolSlice()))
-			for _, v := range inA.Value.AsBoolSlice() {
-				sl.AppendEmpty().SetBool(v)
-			}
-		case attribute.INT64SLICE:
-			sl := dest.PutEmptySlice(key)
-			sl.EnsureCapacity(len(inA.Value.AsInt64Slice()))
-			for _, v := range inA.Value.AsInt64Slice() {
-				sl.AppendEmpty().SetInt(v)
-			}
-		case attribute.FLOAT64SLICE:
-			sl := dest.PutEmptySlice(key)
-			sl.EnsureCapacity(len(inA.Value.AsFloat64Slice()))
-			for _, v := range inA.Value.AsFloat64Slice() {
-				sl.AppendEmpty().SetDouble(v)
-			}
-		case attribute.STRINGSLICE:
-			sl := dest.PutEmptySlice(key)
-			sl.EnsureCapacity(len(inA.Value.AsStringSlice()))
-			for _, v := range inA.Value.AsStringSlice() {
-				sl.AppendEmpty().SetStr(v)
-			}
-		default:
-			panic("unhandled case")
+		CopyAttribute(dest, iter.Attribute())
+	}
+}
+
+func CopyAttribute(dest pcommon.Map, inA attribute.KeyValue) {
+	key := string(inA.Key)
+	switch inA.Value.Type() {
+	case attribute.BOOL:
+		dest.PutBool(key, inA.Value.AsBool())
+	case attribute.INT64:
+		dest.PutInt(key, inA.Value.AsInt64())
+	case attribute.FLOAT64:
+		dest.PutDouble(key, inA.Value.AsFloat64())
+	case attribute.STRING:
+		dest.PutStr(key, inA.Value.AsString())
+	case attribute.BOOLSLICE:
+		sl := dest.PutEmptySlice(key)
+		sl.EnsureCapacity(len(inA.Value.AsBoolSlice()))
+		for _, v := range inA.Value.AsBoolSlice() {
+			sl.AppendEmpty().SetBool(v)
 		}
+	case attribute.INT64SLICE:
+		sl := dest.PutEmptySlice(key)
+		sl.EnsureCapacity(len(inA.Value.AsInt64Slice()))
+		for _, v := range inA.Value.AsInt64Slice() {
+			sl.AppendEmpty().SetInt(v)
+		}
+	case attribute.FLOAT64SLICE:
+		sl := dest.PutEmptySlice(key)
+		sl.EnsureCapacity(len(inA.Value.AsFloat64Slice()))
+		for _, v := range inA.Value.AsFloat64Slice() {
+			sl.AppendEmpty().SetDouble(v)
+		}
+	case attribute.STRINGSLICE:
+		sl := dest.PutEmptySlice(key)
+		sl.EnsureCapacity(len(inA.Value.AsStringSlice()))
+		for _, v := range inA.Value.AsStringSlice() {
+			sl.AppendEmpty().SetStr(v)
+		}
+	default:
+		panic("unhandled case")
 	}
 }