--- a/lightstep/sdk/metric/aggregator/aggregator.go
+++ b/lightstep/sdk/metric/aggregator/aggregator.go
@@ -44,7 +44,7 @@ const (
 	// exemplars require additional synchronization.
 	AlwaysOffKind ExemplarFilterKind = iota
 	AlwaysOnKind
-	TraceBasedKind
+	WhenTracedKind
 )
 
 // Implementation note: this code supports one kind of