--- a/lightstep/sdk/metric/aggregator/histogram/histogram.go
+++ b/lightstep/sdk/metric/aggregator/histogram/histogram.go
@@ -169,3 +169,7 @@ func (Methods[N, Traits]) SubtractSwap(operand, argument *Histogram[N, Traits]) 
 	// which start as delta temporality and thus never subtract.
 	panic("impossible call")
 }
+
+func (Methods[N, Traits]) Exemplars(ptr *Histogram[N, Traits], in []aggregator.WeightedExemplarBits) []aggregator.WeightedExemplarBits {
+	return in
+}