--- a/lightstep/sdk/metric/data/data.go
+++ b/lightstep/sdk/metric/data/data.go
@@ -89,8 +89,9 @@ type (
 		// was performed.
 		End time.Time
 
-		// Exemplars @@@
-		Exemplars []aggregator.ExemplarBits
+		// Exemplars. See the comments on Metrics about re-use
+		// of slices in this struct.
+		Exemplars []aggregator.WeightedExemplarBits
 	}
 )
 
@@ -127,7 +128,20 @@ func (in *Instrument) Reset() {
 }
 
 func resetPoints(ps *[]Point) {
+	for i := range *ps {
+		(*ps)[i].Reset()
+	}
 	(*ps) = (*ps)[0:0:cap((*ps))]
+}
+
+// Reset sets the size of every slice to zero, while maintaining
+// capacity, to support re-use.
+func (p *Point) Reset() {
+	resetExemplars(&p.Exemplars)
+}
+
+func resetExemplars(exs *[]aggregator.WeightedExemplarBits) {
+	(*exs) = (*exs)[0:0:cap((*exs))]
 }
 
 // ReallocateFrom returns the pointer to the next available slice