--- a/lightstep/sdk/metric/exemplar/last.go
+++ b/lightstep/sdk/metric/exemplar/last.go
@@ -15,46 +15,58 @@
 package exemplar
 
 import (
+	"sync"
+
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/aggregation"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/number"
 )
 
 type LastStorage[N number.Any, Traits number.Traits[N], Storage any, Methods aggregator.Methods[N, Storage]] struct {
-	Aggregate Storage
+	aggregate Storage
+
+	lock     sync.Mutex
+	exemplar aggregator.ExemplarBits
 }
 
 type LastMethods[N number.Any, Traits number.Traits[N], Storage any, Methods aggregator.Methods[N, Storage]] struct{}
 
-func (s LastStorage[N, Traits, Storage, Methods]) Kind() aggregation.Kind {
+func (s *LastStorage[N, Traits, Storage, Methods]) Kind() aggregation.Kind {
 	var am Methods
 	return am.Kind()
 }
 
 func (m LastMethods[N, Traits, Storage, Methods]) Init(ptr *LastStorage[N, Traits, Storage, Methods], cfg aggregator.Config) {
 	var am Methods
-	am.Init(&ptr.Aggregate, cfg)
+	am.Init(&ptr.aggregate, cfg)
 }
 
 func (m LastMethods[N, Traits, Storage, Methods]) Update(ptr *LastStorage[N, Traits, Storage, Methods], number N, ex aggregator.ExemplarBits) {
 	var am Methods
-	// TODO: ex
-	am.Update(&ptr.Aggregate, number, ex)
+	if ex.Span == nil {
+		am.Update(&ptr.aggregate, number, ex)
+		return
+	}
+
+	ptr.lock.Lock()
+	defer ptr.lock.Unlock()
+	ptr.exemplar = ex
+	am.Update(&ptr.aggregate, number, ex)
 }
 
 func (m LastMethods[N, Traits, Storage, Methods]) Move(input, output *LastStorage[N, Traits, Storage, Methods]) {
 	var am Methods
-	am.Move(&input.Aggregate, &output.Aggregate)
+	am.Move(&input.aggregate, &output.aggregate)
 }
 
 func (m LastMethods[N, Traits, Storage, Methods]) Merge(input, output *LastStorage[N, Traits, Storage, Methods]) {
 	var am Methods
-	am.Merge(&input.Aggregate, &output.Aggregate)
+	am.Merge(&input.aggregate, &output.aggregate)
 }
 
 func (m LastMethods[N, Traits, Storage, Methods]) Copy(input, output *LastStorage[N, Traits, Storage, Methods]) {
 	var am Methods
-	am.Copy(&input.Aggregate, &output.Aggregate)
+	am.Copy(&input.aggregate, &output.aggregate)
 }
 
 func (m LastMethods[N, Traits, Storage, Methods]) SubtractSwap(operand, argument *LastStorage[N, Traits, Storage, Methods]) {
@@ -77,5 +89,5 @@ func (m LastMethods[N, Traits, Storage, Methods]) Kind() aggregation.Kind {
 
 func (m LastMethods[N, Traits, Storage, Methods]) HasChange(ptr *LastStorage[N, Traits, Storage, Methods]) bool {
 	var am Methods
-	return am.HasChange(&ptr.Aggregate)
+	return am.HasChange(&ptr.aggregate)
 }