**Description:** 
First implementation of the OpenTelemetry metrics SDK exemplar specification. 
This implements the specified exemplar filter mechanism.
This implements two exemplar reservoirs, one named `Last` and one named `Weighted`. 
Last picks a single exemplar, the most recent.
Weighted uses the VarOpt algorithm from https://github.com/lightstep/varopt.

**Link to tracking Issue:**
#575 

**Testing:** WIP

**Documentation:** TODO